+++ packages/example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/realdwg-web-example
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/libredwg-converter@3.4.4
+  - @mlightcad/data-model@1.6.3
+
 ## 0.0.3
 
 ### Patch Changes