+++ packages/libdxfrw-converter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mlightcad/libdxfrw-converter
 
+## 3.4.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/data-model@1.6.3
+
 ## 3.4.2
 
 ### Patch Changes